CommentEntity 생성 및 PostEntity id 이름 부여 및 매핑 - 김찬혁
