--- src/main/java/com/example/projectpicker/post/dto/response/CommentResponseDTO.java
@@ -1,4 +1,0 @@
-package com.example.projectpicker.post.dto.response;
-
-public class CommentResponseDTO {
-}
