--- src/main/java/com/example/projectpicker/postapi/dto/response/PostListResponseDTO.java
@@ -1,18 +1,0 @@
-package com.example.projectpicker.postapi.dto.response;
-
-import com.example.projectpicker.postapi.dto.request.PageRequestDTO;
-import lombok.*;
-
-import java.util.List;
-
-@Setter @Getter @ToString
-@NoArgsConstructor
-@AllArgsConstructor
-@EqualsAndHashCode
-@Builder
-public class PostListResponseDTO {
-
-    private int count;
-    private PageResponseDTO pageInfo;
-    private List<PostResponseDTO> posts;
-}
