--- src/main/java/com/example/projectpicker/postapi/repository/HashTagRepository.java
@@ -1,9 +1,0 @@
-package com.example.projectpicker.postapi.repository;
-
-import com.example.projectpicker.postapi.entity.HashTagEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface HashTagRepository extends JpaRepository<HashTagEntity, String> {
-
-
-}
